# docs/swagger.yaml
definitions:
...
  dto.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
    type: object
...
  dto.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.ThumbnailRegenerateRequest:
    properties:
      height:
        minimum: 1
        type: integer
      timestamp:
        minimum: 0
        type: number
      width:
        minimum: 1
        type: integer
    type: object
  dto.ThumbnailRegenerateResponse:
    properties:
      id:
        type: integer
      message:
        type: string
      thumbnail_info:
        $ref: '#/definitions/entity.ThumbnailInfo'
    type: object
...
  dto.VideoListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/entity.Video'
        type: array
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
...
  dto.VideoYearsResponse:
    properties:
      data:
        items:
          type: integer
        type: array
    type: object
...
  entity.JSONMap:
    additionalProperties: true
    type: object
  entity.Series:
    properties:
      comment:
        $ref: '#/definitions/entity.JSONMap'
      created_at:
        type: string
      first_end_month:
        type: integer
      first_end_year:
        type: integer
      first_month:
        type: integer
      first_year:
        type: integer
      id:
        type: integer
      series_name_file:
        description: '{title} 部分'
        type: string
      subtitles:
        items:
          additionalProperties: true
          type: object
        type: array
      syobocal_title_id:
        type: integer
      syobocal_title_name:
        type: string
      syobocal_title_name_en:
        type: string
      updated_at:
        type: string
    type: object
  entity.ThumbnailInfo:
    properties:
      generated_at:
        type: string
      height:
        type: integer
      width:
        type: integer
    type: object
...
  entity.Video:
    properties:
      channel_id:
        description: Syobocal チャンネルID（ChID）
        type: integer
      created_at:
        type: string
      description:
        type: string
      duration:
        type: number
      episode:
        description: エピソード番号
        type: integer
      file_name:
        type: string
      file_size:
        type: integer
      folder_id:
        description: Folderテーブルへの外部キー
        type: integer
      id:
        type: integer
      is_deleted:
        type: boolean
      jikkyo_comment_count:
        type: integer
      jikkyo_date:
        type: string
      liked:
        type: boolean
      prog_end_time:
        description: 放送終了時刻
        type: string
      prog_start_time:
        description: 放送開始時刻
        type: string
      screenshot_file_path:
        type: string
      series:
        allOf:
        - $ref: '#/definitions/entity.Series'
        description: Series関連付け
      series_id:
        description: Seriesテーブルへの外部キー
        type: integer
      status:
        description: ready, processing, error
        type: string
      subtitle:
        description: エピソードサブタイトル
        type: string
      thumbnail_info:
        $ref: '#/definitions/entity.ThumbnailInfo'
      updated_at:
        type: string
      views:
        type: integer
    type: object
...
info:
  contact: {}
  description: Watch commentPlayer API documentation
  title: commentPlayer API
  version: 1.0.0
paths:
...
  /api/v1/videos:
    get:
      description: ページネーション対応のビデオ一覧を取得します。IDsはコンマ区切りで複数指定可能です
      parameters:
      - description: ビデオID（複数指定可能、コンマ区切り、例：
        in: query
        name: ids
        type: string
      - description: フィルター
        in: query
        name: filterBy
        type: string
      - description: 年フィルター（例：2023）
        in: query
        name: year
        type: integer
      - default: 1
        description: ページ番号
        in: query
        name: page
        type: integer
      - default: 20
        description: 1ページあたりのアイテム数
        in: query
        name: limit
        type: integer
      - default: created_at
        description: ソート対象フィールド
        in: query
        name: sort
        type: string
      - default: desc
        description: ソート順序
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.VideoListResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: ビデオ一覧を取得
      tags:
      - Videos
  /api/v1/videos/{id}:
    get:
      description: 特定のビデオの詳細情報を取得します
      parameters:
      - description: ビデオID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Video'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: ビデオ詳細を取得
      tags:
      - Videos
  /api/v1/videos/{id}/download:
    get:
      description: ビデオファイルをダウンロードします
      parameters:
      - description: ビデオID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: ビデオをダウンロード
      tags:
      - Videos
  /api/v1/videos/{id}/thumbnail/regenerate:
    post:
      description: ビデオのサムネイルを再生成します
      parameters:
      - description: ビデオID
        in: path
        name: id
        required: true
        type: integer
      - description: リクエストボディ
        in: body
        name: body
        schema:
          $ref: '#/definitions/dto.ThumbnailRegenerateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ThumbnailRegenerateResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: サムネイルを再生成
      tags:
      - Videos
  /api/v1/videos/{id}/view:
    post:
      description: ビデオの視聴開始時に呼び出し、views を増加させ watched_history に行を追加します
      parameters:
      - description: ビデオID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: ビデオの視聴を記録
      tags:
      - Videos
  /api/v1/videos/search:
    get:
      description: キーワードでビデオを検索します
      parameters:
      - description: 検索キーワード
        in: query
        name: q
        required: true
        type: string
      - default: 1
        description: ページ番号
        in: query
        name: page
        type: integer
      - default: 20
        description: 1ページあたりのアイテム数
        in: query
        name: limit
        type: integer
      - default: desc
        description: ソート順序
        in: query
        name: order
        type: string
      - description: フィルター
        in: query
        name: filterBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.VideoListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: ビデオを検索
      tags:
      - Videos
  /api/v1/videos/years:
    get:
      description: jikkyo_dateから抽出した年の一覧を降順で返します
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.VideoYearsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: ビデオの年一覧を取得
      tags:
      - Videos
...
schemes:
- http
swagger: "2.0"